+++ test/e2e/tests/fixes/143.mjs
@@ -13,6 +13,11 @@ const testSteps = [
 					values: ['a', 'a', 'a', 'b', 'b', 'b', 'c', 'c', 'c']
 				},
 				{
+					name: 'Letters2',
+					type: 'dimension',
+					values: ['a', 'a', 'a', 'a', 'a', 'b', 'b', 'b', 'b']
+				},
+				{
 					name: 'Val',
 					type: 'measure',
 					values: [3, 5, 4, 3 + 1, 5 + 1, 4 + 1, 3 + 2, 5 + 2, 4 + 2]
@@ -24,19 +29,55 @@ const testSteps = [
 			data,
 			config: {
 				y: 'Colors',
-				x: 'Val'
+				x: 'Val',
+				split: true
 			}
 		})
 	},
 	(chart) =>
 		chart.animate(
 			{
-				y: 'Val',
-				x: 'Letters',
-				split: true
+				y: ['Letters2', 'Val'],
+				x: 'Letters'
 			},
 			{ regroupStrategy: 'drilldown' }
-		)
+		),
+	(chart) =>
+		chart.animate(
+			{
+				y: ['Letters2', 'Colors', 'Val']
+			},
+			{ regroupStrategy: 'drilldown' }
+		),
+	(chart) =>
+		chart.animate({
+			config: {
+				y: 'Colors',
+				x: 'Val'
+			}
+		}),
+	(chart) =>
+		chart.animate({
+			config: {
+				y: ['Letters2', 'Val'],
+				x: 'Letters'
+			}
+		}),
+	(chart) =>
+		chart.animate({
+			y: ['Letters2', 'Val']
+		}),
+	(chart) =>
+		chart.animate({
+			y: ['Letters2', 'Colors', 'Val']
+		}),
+	(chart) =>
+		chart.animate({
+			config: {
+				y: ['Letters2', 'Colors'],
+				x: 'Val'
+			}
+		})
 ]
 
 export default testSteps