+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Fixed
 
 - Fix align on fake-split charts
+- Drilldown on split chart is fade.
+- Fix aggregate on split chart.
 - Do not interpolate hiding/showing legend
 - Fix aggregator interface for 'set' channel parameter:
   - From now not accepted the same dimension on the same channel.